+++ src/components/common/Filter/SemesterDropdown.tsx
@@ -2,27 +2,29 @@ import { Box, FormControl, InputLabel, MenuItem, Select, styled } from '@mui/mat
 import { useEffect } from 'react';
 import { useDispatch, useSelector } from 'react-redux';
 import { setSemester } from 'src/redux/slices/filter';
+import { generateSemesters, currentYear } from 'src/utils/semesterAutoGenerate';
+// const SEMESTERS = [
+//   '전체',
 
-const SEMESTERS = [
-  '전체',
-
-  '2020-01',
-  '2020-02',
-  '2021-01',
-  '2021-02',
-  '2022-01',
-  '2022-02',
-  '2023-01',
-  '2023-02',
-  '2024-01',
-  '2024-02',
-  '2025-01',
-  '2025-02',
-  '2026-01',
-  '2026-02',
-];
+//   '2020-01',
+//   '2020-02',
+//   '2021-01',
+//   '2021-02',
+//   '2022-01',
+//   '2022-02',
+//   '2023-01',
+//   '2023-02',
+//   '2024-01',
+//   '2024-02',
+//   '2025-01',
+//   '2025-02',
+//   '2026-01',
+//   '2026-02',
+// ];
 
 export default function SemesterDropdown() {
+  const SEMESTERS = generateSemesters(currentYear);
+
   const semester = useSelector((state) => state.filter.semester);
   const dispatch = useDispatch();
