+++ src/components/common/Select/SemesterMagicianSelect.tsx
@@ -2,26 +2,27 @@ import { FormControl, InputLabel, MenuItem, Select } from '@mui/material';
 import { ErrorMessage, Field } from 'formik';
 import { useSelector } from 'react-redux';
 import { CATEGORYID, SEMESTER } from 'src/assets/data/fields';
-
+import { generateSemesters, currentYear } from 'src/utils/semesterAutoGenerate';
 export default function SemesterMagicianSelect({ semester, setSemester }) {
-  const semesters = [
-    '2020-01',
-    '2020-02',
-    '2021-01',
-    '2021-02',
-    '2022-01',
-    '2022-02',
-    '2023-01',
-    '2023-02',
-    '2024-01',
-    '2024-02',
-    '2025-01',
-    '2025-02',
-    '2026-01',
-    '2026-02',
-    '2027-01',
-    '2027-02',
-  ];
+  // const semesters = [
+  //   '2020-01',
+  //   '2020-02',
+  //   '2021-01',
+  //   '2021-02',
+  //   '2022-01',
+  //   '2022-02',
+  //   '2023-01',
+  //   '2023-02',
+  //   '2024-01',
+  //   '2024-02',
+  //   '2025-01',
+  //   '2025-02',
+  //   '2026-01',
+  //   '2026-02',
+  //   '2027-01',
+  //   '2027-02',
+  // ];
+  const semesters = generateSemesters(currentYear);
 
   return (
     <Select
